requirements.txt
feedparser
django-conneg
html2text
redis>=2.4.9
sharepoint
django-maintenancemode
django-dirtyfields
python-dateutil
raven
django-pipeline
requests
python3-ldap